Move processing functions to oemof_b3.tools
